feat(SnapDropZone): add valid highlight color option
The new Valid Highlight Color will be used when a valid object is
hovering over an already highlighted snap drop zone. For instance
if the Highlight Always Active option is on, it will use the
Highlight Color (which can be set to a different colour) and then
when a valid object passes into the Snap Drop Zone the Valid
Highlight Color will be used.

This also fixes an issue with the OnEnter event as it wasn't
previously being emitted if the Highlight Always On was set to
true, until the object was snapped and then unsnapped.