chore(Structure): support new XR naming convention in 2017.2
Unity 2017.2 has changed all of the VR packages and VR class names
and renamed them to XR. To ensure future proofing, all of the
references to VR classes have been updated XR and then script
define symbols for any Unity version older than 2017.2 create
class aliases from the old VR convention to the new XR convention.

The VRSettings.renderScale method has been completely renamed so
it's not possible to alias it. In this case a new helper method
has been created in SharedMethods that allows the retrieval of
the renderScale and another method which sets it.