feat(Interaction): ability to cancel haptics on interaction end
A new setting has been added to the Interact Haptics script that lets
the current haptic rumble be cancelled if the interaction type ends.

For example, the touch interaction can cause a long rumble to occur
but as soon as an untouch happens the rumble can now be stopped
immediately.

A small change to the Interact Near Touch script was also made to
ensure that the event process when going from near touch to touch went:

 * Near Touch
 * Near Untouch
 * Touch

Whereas previously it was doing:

 * Near Touch
 * Touch
 * Near Untouch