fix(Interaction): ensure controller highlighter uses correct aliases
The Controller Highlighter script was not using the correct model alias
or the correct actual controller model if one was specified. It would
also fail if the SDK Manager was switched to another SDK.

This has been fixed by ensuring the references to the relevant objects
are set accordingly and not overridden incorrectly.