feat(Interaction): provide custom highlighter to interactable object
  > **Breaking Changes**

The Interactable Object script now has an additional parameter that
allows injecting a Base Highlighter for the Interactable Object to
use.

This means that the Highlighter does not need to be on the same
GameObject as the Interactable Object and can also be swapped out
at runtime for a different highlighter.

To enable this to work, the Highlighters now can accept new
GameObject parameter in the Initialise method. This GameObject
parameter is the second parameter pushing the `options` parameter
to the 3rd param which will cause a breaking change for anyone
using the existing code.