fix(Pointers): turn off renderer when either script is disabled
There was an issue where if the Pointer was active in the scene but
either the Pointer script or the Pointer Renderer script were disabled
then the Pointer Renderer would stay visible on the screen.

This fix ensures that when the Pointer script is disabled it
automatically calls `Toggle(false)` to turn off the pointer and also
checks the state of the attached Pointer Renderer and if it is disabled
the `Toggle(false)` is also called to ensure the pointer is disabled.