fix(SnapDropZone): prioritise joint removal on highlight object
If the interactable object that is being used as the highlight object
template has a joint on it, then it would cause issues due to the
rigidbody attempting to be removed first which would then fail because
a Joint was present. This would cause the snap drop zone highlight
object to keep it's rigidbody and fall will the physics gravity when
the scene started.