feat(Locomotion): add tunnel overlay camera effect for comfort options
The Tunnel Overlay script creates a texture overlay that can increase
in it's radius towards the centre of the screen to reduce the perceived
field of view.

This can be triggered based on the movement and rotation of the play
area to reduce nausea with artificial locomotion. The tunnel effect
can also be applied when stationary to simulate effects such as damage
being taken.

Thanks to @reznovVR for the original port of the excellent
UnityVRTunnelling repo by @sixways:

https://github.com/SixWays/UnityVrTunnelling