chore(Structure): reorganise interaction scripts into sub directories
The Interactions scripts have been reorganised into two new sub
directories, any script that is to be attached to an interacting object
like a controller is now inside the `Interactors` directory and any
object that is to be interacted with like an Interactable Object is
now inside the `Interactables` directory.

The auto gen documentation has been updated to ensure this new split
works in the documentation.