feat(Interaction): add physics rotator controllable
The Physics Rotator controllable creates an interactable object
that is attached to a hinge joint and can be rotated around
the hinge joint upon grab and movement of the controller.

The joint position is determined by a specified transform which
removes any magic setup and allows for greater control.

The Door, Chest, Lever and Spring Lever 3D Controls have now been
deprecated as they can be recreated with the new physics rotator
controllable.