feat(Locomotion): apply play area parent position to teleporter
The Height Adjust Teleport script now has a new option that will also
apply the play area parent position as an offset when teleporting.

This fixes issues with headsets that aren't room scale aware and
therefore cannot determine where the floor is so will always send the
HMD view to the point of teleport. This would cause the user to
be teleported into the floor.

This new feature allows for the play area camera rig to be placed
in a parent GameObject and that parent GameObject to have a Y position
set on it to be used as the default standing height of the HMD view.