feat(Structure): provide custom data types to convey better meaning
  > **Breaking Changes**

In a number of scripts, standard data types were being used but created
confusion in meaning, such as using a `Vector2` for a min/max range as
this would then display in the editor as `x` and `y` values, with `x`
being minimum and the `y` being maximum.

This has now been replaced with a new `Limits2D` data type that
displays the same 2 float input boxes as a `Vector2` but instead uses
the labels of `min` and `max` which makes a lot more sense.

The `Limits2D` data type can also have a new custom attribute applied
to it called `[MinMaxRange]` which works in a similar way to `[Range]`
but it provides 2 slider points that can be dragged to set a minimum
and maximum value for the control. It should also work with a Vector2
if required.

The Axis Scale Grab Action has also been updated to use the new
Vector3State data type when determining which axes should be locked
for scaling. This is in place of the previous 3 separate boolean
parameters.