fix(Pointers): use appropriate tip for determining pointer origin
Previously the Pointer and UIPointer would use the transform of the
controller for the pointer origin and forward, but for SteamVR due
to different controllers being at different orientations it would
use a custom origin via the `GenerateControllerPointerOrigin` method.
This is not required as the `tip/attach` GameObject in a SteamVR
controller is already at the correct orientation for determining the
appropriate pointer origin.

This fix now ensures this tip is used for determining the pointer
origin and the `GenerateControllerPointerOrigin` class has been
deprecated as no other SDKs used it except for SteamVR and SteamVR
already has a built in mechanism for dealing with this complication.

The Pointer and UIPointer also have a new `Attached To` parameter
which can be used to determine which GameObject the pointer is actually
attached to, which helps when determining the pointer origin as if
the pointer is not attached to a controller.

The `Controller` parameter has also been deprecated and brought into
line with other scripts that correctly name it `Controller Events`.