fix(Internal): ensure controller ready registration is always done
The SDK Controller Ready script was only attempting to register the
controller ready check if a valid controller reference was found.

However, with SteamVR and possibly other SDKs the controller is not
actually ready at start up and takes a few moments before it's ready
meaning the controller reference is not valid.

The registration of the controller ready checks should always be
done anyway and not only if a valid controller is ready as the
SDK event is emitted when the actual SDK reports the controller as
being ready.