feat(SDK): add hand axis guides to simulator
The Simulator hands now have an axis guide to show the axis in which
the hands will be moving.

Also, added a parameter to set the speed of the mouse cursor when it
is in locked mode as it was moving too slowly.

All private methods have also been made protected now if the script
requires extending.

Only one example scene has been updated with the new Simulator
CameraRig prefab as the rest of the example scenes have become detached
from the prefab due to the name change.