feat(Locomotion): option to add skybox texture to tunnel overlay cage
The view in which the Tunnel Overlay fades out to can now have a
Cubemap texture provided to display as a static skybox to provide a
better point of reference.

The Effect Color can also be used to tint the skybox texture if
required.

Thanks to @jwnjwn for the work on this!