feat(SDK): utilise underlying sdk to retrieve headset type
Previously, the `VRTK_DeviceFinder.GetHeadsetType()` would rely on the
built in Unity methods for determining the headset type, however this
was unsustainable as it didn't report generic enough information and
would mean the list of headset types would grow and grow and ultimately
not provide useful information.

This has now been changed by each SDK now has a `GetHeadsetType` method
that reports a simpler to understand headset type string, so in the
case of the vive, it does not report all of the variants, just reports
"htcvive" and similarly with the rift.

This string is then used to return a new, cleaner enum from the
`VRTK_DeviceFinder.GetHeadsetType()` method and the enum is stored
in the SDK_BaseHeadset class as it's more appropriate.

The usage of `GetHeadsetType` to determine which controller is
connected in the SDK_SteamVRController class has been removed and
now the render model of the connected controller is used to determine
the controller type as this will always be accurate from what OpenVR
is reporting.

When using the Unity SDK, the headset is still retrieved in the old
way as Unity doesn't provide enough decent information about the
connected headset to know which device is actually connected. It
does this by scraping the model and device name.