feat(Interaction): remove highlighting duty from interactable obejct
The Interactable Object is no longer responsible for causing the object
to highlight. All of the highlighting code within the Interactable
Object script has now been removed or deprecated. The Interact
Object Highlight script now takes full control of highlighting an
object and it has been improved so it can listen for the events on an
interactable object and affect the highlighting of a different object.