feat(Input): ability to switch script input button based on sdk type
A new SDK Input Override script has been added that allows the input
button on the numerous scripts to be changed at runtime based on the
SDK being used or the controller type connected.

This script is in place of the proposed signal system which will be
more robust and more versatile than the sdk input override solution.

However, due to the way controller events still work, this is the
only viable solution right now. Hopefully, this script can be
deprecated in the future when/if the new signal system is implemented.