fix(Pointers): clear cached attach point on sdk change
The cached attach point on the Base Pointer Renderer was not being
cleared when the SDK was switched so the transform follow script was
not updating to track the new hand type.

This has been fixed by clearing the cache in the `OnEnable` method
and registering the script with the SDKManager toggle method.