fix(Interaction): allow controllables to be moved without breaking
The Controllables, specifically the Physics based ones were not working
correctly if they were moved in the scene and then disabled/re-enabled
due to how the position was being calculated based on the original
scene position being used to determine the joint set up. The original
position was being re-set every time the object was enabled which was
partially the issue, but also the original position was not being
taken into consideration when updating the transform local position.