feat(SDKManager): allow delayed assignment of behaviour toggles
The `AddBehaviourToToggleOnLoadedSetupChange` method can now be called
via a static method called
`AttemptAddBehaviourToToggleOnLoadedSetupChange` which will attempt to
add the behaviour if the SDK Manager instance is valid, if it's not
valid then it's added to a HashSet and can be manually processed at
a later time with `ProcessDelayedToggleBehaviours`.

This is useful if the SDK Manager is in an additively loaded scene
but other scripts in the original scene are attempting to register
their behaviour in the `Awake` method.

This will make it easier to additively load in a constructor scene
and the SDK Manager will still work correctly.

A bunch of other SDK Manager public methods have also had static
counterparts added for them, making it easier to call methods without
always having to check if the SDK Manager instance is valid.