fix(Structure): clone HashSet when iterating to prevent change errors
There was an issue when using a HashSet and a foreach loop that if the
HashSet contents had changed during the iteration then an error would
be thrown.

The solution is to clone the HashSet before iterating over the clone
so any changes to the original HashSet don't affect the iteration.