feat(SDK): add native Unity Windows Mixed Reality support
Support for the Unity Windows Mixed Reality WSA namespace to allow
for Windows Mixed Reality headsets and controllers to be used within
VRTK.

A number of reflection methods have been abstracted into the
Shared Methods class to remove errors when building on the Unity WSA
platform with the .NET scripting backend. These errors are caused
by the alternative version of .NET and differing ways of handling
reflection code.

Controller visualisation is provided by a 3rd party repository and
instructions of how to use are provided in the WindowsMR README.md
file.

A massive thanks to @tomwim for all his hard work on this and thanks
to @DDReaper for continued support along the way.