fix(Examples): use final scene index for constructor scene
Previously, the constructor scene was injected into the
AdditiveSceneLoader component as a `Object` type and then the `name`
of this object was used to identify the constructor scene which worked
fine when running in the Unity Editor but would fail in a build.

This is most likely because the `Object` type is not being injected
in a build and only works due to quirks in the editor.

The solution is now to ensure the final scene in Build Settings is the
constructor scene and then simply to use that index in the
AdditiveSceneLoader component and remove the ability to inject a scene.