fix(Examples): ensure constructor scene is available in build settings
Previously, the constructor scene was removed from being injected into
the example scenes as it wasn't working in a build, instead the
constructor scene was required to be the last scene in the build
settings and was loaded in by index.

This only worked if using the project settings from the VRTK repo but
would not work if VRTK was imported into another project.

To remedy this, the scene injection is now being used again and the
last scene index is being used as a fallback. If no scene is in the
build settings then the script will attempt to add it as the last
scene automatically.