fix(Presence): provide custom boxcast to ensure correct settings used
The Body Physics script was doing a BoxCast but always just using the
default `Physics.BoxCast` which meant it never took into consideration
the custom raycast settings being applied.

The CustomRaycast script now contains a CustomBoxCast method that the
Body Physics script is now using which brings it in line with other
raycast types.