chore(structure): replace submodule with newly named repo
The VRTK.Unity.Core submodule has been renamed to Zinnia and this new
submodule has been included.

Along with the rename, all of the prefabs from Zinnia have been removed
as they are to be added to this VRTK repo. But before that it means all
of the existing examples have also been removed as they will be newly
created in a different style in an upcoming commit.

The relevant docs have also been updated to bring a level of
consistency to the documentation across various repositories.