fix(Helpers): reduce tooltip public field options
The public field options on the Tooltip prefab have been vastly reduced
so instead of having helper fields for common tooltip items, these
fields are now expected to be set on the underlying tooltip component.

For example, the text is actually a UI Text element nested in the
prefab and to change the text and appearance of this text then it is
better to just edit the UI Text element component directly.