feat(examples): add tooltip popup
The Tooltip prefab has been added to the tools drawer in the warehouse
and is on a pointing query so the tooltip only becomes visible when
the HMD is pointing at the tools drawer.

It's a simple showcase of how to use tooltips and how to use the
pointing query.