feat(examples): add interactable well
The Interactable Well utilizes a RotationalJointDrive to simulate
turning a handle to lower and raise a bucket.

The lowering and raising of the bucket is a simple animation scrub
that is achieved with the helper script AnimatorScrubber which
takes an Animator and an AnimationName and scrubs to the normalized
position of the animation timeline.

This is then hooked into the normalized value of the drive to ensure
it tracks the rotation correctly.