fix(examples): ensure locomotion state is enabled after falling
There was an issue with the locomotion state not being re-enabled when
throwing yourself after climbing and falling due to the timers calling
cancel but if the timer had already completed then the cancel event
would never emit and the locomotion would never be re-enabled.

There was also another issue where the climbing prefab would disable
the entire Movement GameObject meaning the movement state logic
may not run in time. This logic has now been moved out of the
Movement GameObject so it is never disabled.