fix(Interactions): set rigidbody maxAngularVelocity to infinity
Unity sets a rigidbody maximum angular velocity to 7 mps by default
which is acceptable for most uses, however with tracking rigidbodies
to follow the angular changes of a posed controller then this max
velocity is not enough.

This manifests into an issue with using rigidbody tracking on the
object follow as the target object slowly rotating to keep up with
the source object.

The fix is to set the Rigidbody.maxAngularVelocity parameter to
infinity. This cannot be done on the Rigidbody inspector component
as Unity do not expose this parameter for some reason. Instead of
adding a new component, the BehaviourEnabledObserver has been added
to the Interactables that listens for when the InteractableFacade
is enabled and then sets the maxAngularVelocity on all of the relevant
rigidbodies.