fix(examples): prevent wind mill and floor from fading view
Due to an issue where if the headset starts in the floor it will cause
the view to fade due to the CollisionFader on the headset, so it's
easier to just exclude the floor from being able to fade the headset
as it's not likely users will be putting their head in the floor.

The Wind Mill also has an issue where the climbable mesh is a convex
collider which means the whole section even where there is no visible
mesh is considered a collider that will fade the headset, this means
that simply standing underneath the wind mill would cause the view to
fade. For the simplicity of the scene, it's easier to just ignore
the wind mill from fading the view. It also shows how it's possible
to exclude climbable elements from fading the view.