fix(examples): ensure axis rotation uses headset as rotation origin
The axis rotation locomotion now uses the new rotation origin property
on the TransformEulerRotationMutator to ensure when the play area is
rotated that it rotates around where the user is standing (i.e. the
headset location).

Because the rotation origin must be a child of the target, a new
object follower proxy has been set up that is a child of the playarea
alias and simply follows the position of the headset alias.