fix(CameraRig): use appropriate targets for controlling simulator
The SimulatedCameraRig didn't correctly represent how an actual
VR CameraRig is set up. The need to inject the TrackedAlias to
move the simulator rig around is incorrect and in fact the simulator
should not be allowing the movement of the CameraRig (play area)
as this is not something that happens with actual VR rigs. Instead
a player can walk around their play area but this is just moving their
headset and controller tracked avatars within the play area and then
the play area is moved via other mechanisms.

This fix removes the need to inject a TrackedAlias into the Simulator
and now the movement only controls the position of the headset within
the play area.

The speed that the simulated avatar objects can now also be set on the
SimulatorFacade.