fix(Interactions): clear stop grab publisher collection on stop grab
The ActiveCollisionPublisher for the StopGrabbing action was not being
cleared when a grab was stopped. This meant that if the grabbed object
was swapped between interactors then the previous interactor would
still have the object in its StopGrabbing Publisher and therefore
calling the deactivate action on that interactor (e.g. releasing the
grab button) would cause the interactor to publish the StopGrabbing
data for the grabbed object.

This fix also tidies up the way the Interactable calls methods on the
Interactor, it now goes through the Facade rather than having intrinsic
knowledge of the underlying Interactor Configurators.

This also provides the fix of clearing the ActiveCollisions when an
Ungrab occurs.