feat(examples): ignore collisions between interactors
As Interactors have a CollisionTracker on them it can cause an issue
where the CollisionTracker on one interactor notifies the
CollisionTracker on the other interactor of a collision between each
other and potentially create an unwanted cyclic reference.

This just adds a CollisionIgnorer to remove the collisions between
the left and right Interactor.