fix(Pointers): ensure correct order for pointer actions
There was a race condition where the activation action could be
registered before the selection action meaning doing things like
select on release would not work as the activation action was
being called before the selection action so the pointer would be
disabled before it could select anything.

This fix moves around some of the registration logic to ensure
the correct order for actions.