feat(examples): windows mixed reality style pointer activate/select
A new Axis Control for pointer activation/selection has been added to
the example scene to mimic the way WMR handles pointer activation and
selection by only activating the pointer if the thumbstick axis is
pressed forward and then is only deactivated when the thumbstick
Vector2 axis falls underneath a certain threshold (e.g. returns to
the centre).

This example is all done with wiring up Action events and requires no
custom coding. The new feature can be toggled on/off on the Options
Menu Station under the Button/Axis Control lever in the Pointer Options
section.