fix(Interactions): apply correct kinematic state on follow action
The GrabInteractableFollowAction has been changed so it requires the
specific kinematic state of the containing Rigidbody to be specified as
a parameter for when the follow is active and when the follow is
inactive.

Previously, it would always force the Rigidbody kinematic state to
`true` when a TransformFollow occured but never change the kinematic
state under any other circumstance. It would also always set the
kinematic state back to `false` when the follow finished during
TransformFollow but never did anything with the state under any
other circumstance.

This created an issue as if the object was initially kinematic then
on stop following it would always force it back to being non kinematic
or if the object was kinematic and it was a RigidybodyFollow then as
that follow mechanism requires the Rigidbody to be non kinematic then
it would simply not work.

This new solution means each individual object can be specified what
it expects the Rigidbody kinematic state to be in when a follow starts
and when it ends. It's also possible to change the expected states
during a follow.

So an object may start as being non-kinematic when the follow ends
but during a follow it can be updated so it will stay being kinematic
when the follow ends.

This change will cause some breaking issues for any unpacked prefabs
as the default states may not be correct for the specified follow type.