fix(examples): ensure apple prefab is not kinematic on release
The apple prefab was not becoming non-kinematic when it was picked
and thrown due to a new setting on the Follow mechanic that by default
inherits the default kinematic state of the interactable rigidbody.

Because the apple prefab starts off as kinematic, this inheritance was
making the ungrabbed state of the apple become kinematic as well so the
apple could not be thrown. This fix just stops the auto inheritance
taking place so the actual state can be set manually.