fix(Interactions): track all active collisions on interactable
The Interactable now tracks all objects that are colliding with it via
a GameObject Observable List regardless of whether the object is a
valid interactor. This collisions list can be used to determine what
non interactor objects are touching the interactable for any number
of reasons, one specific reason is to ensure a SnapZone can determine
which interactables are touching which SnapZones without needing to
have an implicit reference on the Interactable.

Common functionality across the different Interactable prefabs has
also been moved into a Common prefab so if generic functionality
changes then it is automatically reflected across all Interactable
prefabs.