fix(Interactions): prevent object activating multiple snapzones
There was nothing preventing an Interactable activating multiple
SnapZones at the same time, but this raised an issue of when the object
was released, then which zone should it snap to? It would snap to the
first zone it collided with but it would also run the snap logic on the
other SnapZone that it had activated.

An object can only activate one SnapZone at a time, even though
multiple objects can enter a SnapZone and all be potentially snappable
to that zone. However, the first object entering is the only object
that actually activates the zone.

This has been accomplished by utilizing the new collision tracking on
the Interactable and determining whether the interactable touching
the current SnapZone is also colliding with another SnapZone. If it is
then an event is registered for when the interactable leaves the
current activated SnapZone to try and then reactivate any other
SnapZones that it also may be touching.