fix(Locomotion): prevent teleport snap to floor onto trigger colliders
The SnapToFloor prefab was set up so the SurfaceLocator would cast a
ray and consider a trigger collider a valid floor. This makes little
sense as a trigger collider is usually used to determine a volume that
can be entered and passed through. The fix is to add a custom
PhysicsCast to the SnapToFloor prefab that has the setting to ignore
trigger colliders for the RayCast. This can then still be disabled
if required.