build: use Zinnia as dependency package
The content of the dependency Zinnia was copied directly into this
project via a git submodule. Keeping dependencies in the project
only is necessary in environments that don't offer the idea of a
package and a package manager. With Unity nowadays coming with the
Unity Package Manager (UPM) VRTK can now become a package,
referencing its needed dependency Zinnia as a package dependency.

BREAKING CHANGE: VRTK is now a UPM package and does no longer
directly include Zinnia, instead referencing it as a package
dependency. Consumers of VRTK will have to follow the added steps in
the README to include VRTK in their projects.