feat(structure): create an example Unity project
The VRTK repo is now simply an example Unity project that includes the
relevant dependencies of the VRTK prefabs to showcase how they can
be used to build something.

It removes the ability to add VRTK to a project via a package.json as
this is now expected to be done by adding the VRTK.Prefabs repo.