feat(examples): add farm example scene disclaimer
A disclaimer has been added to the farm example scene explaining the
scene is a showcase of how VRTK v4 can be used and not necessarily
best practice of how things should be achieved. This disclaimer is
displayed in the Game view via a screen space UI so it will not be
visible in VR and it also disappears after a short amount of time.