fix(examples): remove unnecessary Vector2ComponentExtractor
The Vector2ComponentExtractor component is unnecessary because it is
providing the same functionality as the Vector2ToFloat component.

So the Vector2ComponentExtractor component has been removed from the
scene where it was being used in the:
`VRTK Prefabs -> Controller Actions -> RightController -> Pointers ->
SimulatedRotator -> OutputAxes -> SimulatedHorizontal/Vertical`

for extracting the Vector2 element into a float value. This can simply
be achieved by converting the Vector2 to a float and defining the
coordinate to convert to the output float.