feat(examples): update example project to use tilia packages
The Farm Yard Example project has been updated to use the latest Tilia
packages but is still in development and is missing a number of
features from the previous release that used the deprecated VRTK
Prefabs package.

The current missing features are:

* Locomotion
  * Axis Movement
  * Step Multiplier
  * Drag World
  * Climbing
* Pointers
  * PlayArea Boundary Cursor
  * Distance Grab
* Snap Zone Scarecrow
* Cloneamatic Machine
* Car Lift in warehouse
* Interactable Carry Ladder
* Barn Zip Wire

These features will be added in due course.