feat(project): bump to use unity 2019.4 lts version
The example project has been updated so it runs by default on Unity
2019.4 which is the LTS version for Unity 2019. This is because
Unity struggles to update the project from Unity 2018 to Unity 2019
and causes a bunch of malimbe errors as the compiler gets confused.

It is still being held back from a 2020 LTS update due to the legacy
camera rig and legacy input system, but it may be upgraded in due
course.