feat(examples): add distance grabber option to menu
The distance grabber option has been added to the menu instead of
being a double click of the grab button as it was cumbersome and
wasn't obvious. The logic is still in the scene but just disabled
in case anyone wants to see how to do that sort of thing.